in database use number for type of blockNumber
Suppose we have two entries in DB:
- 1/0xabc/1/9
- 1/0xabc/1/10

If I asked the database for the latest value of `1/0xabc/1` it should
return me `1/0xabc/1/10` but it returns `1/0xabc/1/9`. This is because
according to LevelDB 9 is greater than 10.

This bug has been solved by padding the block number such that 9 is 009
and 10 is 010. Actually, padding can be done for more than three digits.
Here it is just an example.

Returning id as `1/0xabc/1/009` can create confusion. Also, if tomorrow
the database decides to pad it upto four digits (`1/0xabc/1/0009`) the
upper levels should not care. Therefore, to abstract the implementation
detail we have made the type of block number as "number".