transfer close to mints should now be working and rename config types
This commit fixes a bug that makes the crawler skip NFT transfers that
are close to their mints. This was a timing issue where the
transfers were processed before mints were processed.

Along with implementing the fix for the above problem I have also
refactored the code that should make the crawling faster. More NFTs are
now processed in parallel.

This commit also includes a rename in the config.

`config.step.block` and `config.step.contract` are now called
`config.getLogsBlockSpanSize` and `config.getLogsAddressSize`. The
intention is to have variable names that explain their use.

A comment has also been added to explain these config variables.