add localstorage
LocalStorage is similar to the local-storage present in browsers.

It is a simple key-value database powered by LevelDB.

Each strategy can have its own localstorage and use it to store any data
other than tracks.

Strategies are free to use it however they wish.

The purpose of localstorage is to store intermediate data. Here tracks
are the final data.

For example, a strategy like SoundProtocol needs to keep track of
contracts which can mint NFTs and listen for updates on them.
LocalStorage can be used here.

Earlier, we created a JSON file to store the data required by
SoundProtocol. LocalStorage replaces that JSON file as it is more
generic and scalable.