add lens strategy and refactor core structure of strategies
Major changes:

- The Strategy class has been changed. All strategies implement this
class.
- Introduce a concept of *components*.
- Add Lens strategy.
- Refactor Sound Protocol strategy according to the new strucutre.
- Introduce multi-chain strategies.
- Changes in daemon.ts, neume.ts to run the new refactored strategies.

What are components?

Each strategy is a class and it can define functions. What if we want
to share some functions among strategies. The answer to do that is
components.

For example, `ethGetLogs` is a component that can be used to fetch
logs.

Difference between components and utils:

Utils are simple functions which can be called from anywhere.
Components can only be called from inside strategies. We use the `this`
keyword to enforce it. TypeScript helps here by defining the type of
`this`.

Why can't we have utils instead of components?

Components are aware that they will be called from a strategy.
Therefore, they can use the `this` keyword to fetch information about
the strategy and act upon it.

For example, the `ethGetLogs` component can use
`this.chain` to get the chain of the strategy and fetch logs from the
appropriate chain.

If `ethGetLogs` was a simple function we would have to pass the chain
as an argument. It is okay if it is just chain but functions often
need more information than just chain. Passing 10 parameters is not
ideal and problematic from a refactor point of view.

About `handleTransfer` component:

It is a component that can be used by any strategy to listen for
ERC721 transfer events. The component is heavily opnionated.